split target function into separate file
